requirements.txt
# LLM & Agent
langgraph>=0.2.60
langgraph-checkpoint-postgres>=2.0.0
langchain-openai>=0.2.14
langchain-core>=0.3.28
openai>=1.60.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0

# Database
asyncpg>=0.30.0
psycopg[binary]>=3.2.0

# HTTP Client
httpx>=0.27.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Utilities
python-dotenv>=1.0.0